Update README.md
At Arbor Assets, we contribute financial and knowledge-based resources to assist poor, rural communities in Honduras and Nicaragua. Founded in 2009, we are a U.S. 501(c)3 non-profit organization based out of Lincoln, Nebraska. We are principally a micro-lending company, though we have also made strategic, capacity-expanding grants to our partner organizations, usually in terms of additional education, equipment, or materials.